requirements.txt
asgiref==3.5.2
Django==4.0.7
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
dnspython==2.2.1
inflection==0.5.1
PyJWT==2.5.0
pymongo==4.2.0
pytz==2022.2.1
sqlparse==0.2.4
tzdata==2022.2
